Update drivebase_odom_publisher.cpp
v_lin (line 107) and v_ang (line 108) are using tread distance, not speed, so we need to add a time component. 

Updated some constants and added some comments. The turning accuracy is sensitive to wheel base.  Will need to experimentally get it right, because the ideal value is not best for real world performance.  This value currently is off by about 5% when turning.